Minor improvement
